requirements.txt
streamlit==1.45.0
pandas==2.3.0
python-dateutil==2.9.0.post0
pypdf==5.6.0
python-docx==1.2.0
openpyxl==3.1.5
pytest==8.3.5
